% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitR-package.R
\docType{package}
\name{zeitR-package}
\alias{zeitR}
\alias{zeitR-package}
\title{zeitR: Actigraphy Data Parsing and Analysis for R}
\description{
Provides tools for importing, parsing, and analysing raw actigraphy data.
Supports common device file formats, computes standard rest-activity
variables (IS, IV, RA, L5, M10), and assembles tidy data frames ready for
downstream chronobiological analysis. Designed to complement \code{slumbR} in
the Circadia Lab ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zeitr.circadia-lab.uk}
  \item \url{https://github.com/circadia-bio/zeitR}
  \item Report bugs at \url{https://github.com/circadia-bio/zeitR/issues}
}

}
\author{
\strong{Maintainer}: Lucas França \email{lucas.franca@northumbria.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})

Authors:
\itemize{
  \item Lucas França \email{lucas.franca@northumbria.ac.uk} (\href{https://orcid.org/0000-0003-0853-1319}{ORCID})
  \item Mario Leocadio-Miguel \email{mario.miguel@northumbria.ac.uk} (\href{https://orcid.org/0000-0002-7248-3529}{ORCID})
  \item Julia Ribeiro da Silva Vallim (\href{https://orcid.org/0000-0001-8708-8479}{ORCID})
}

}
\keyword{internal}
